* Relocated new classes * Fixed CompositeStrategy refactor
