+++ fabric/forgero-fabric-core/src/main/java/com/sigmundgranaas/forgero/fabric/initialization/ForgeroPreInit.java
@@ -8,12 +8,12 @@ import java.io.InputStreamReader;
 import com.google.gson.Gson;
 import com.google.gson.stream.JsonReader;
 import com.sigmundgranaas.forgero.core.Forgero;
-import com.sigmundgranaas.forgero.core.model.v1.match.PredicateFactory;
-import com.sigmundgranaas.forgero.core.model.v1.match.builders.string.StringIdentifierBuilder;
-import com.sigmundgranaas.forgero.core.model.v1.match.builders.string.StringModelBuilder;
-import com.sigmundgranaas.forgero.core.model.v1.match.builders.string.StringNameBuilder;
-import com.sigmundgranaas.forgero.core.model.v1.match.builders.string.StringSlotBuilder;
-import com.sigmundgranaas.forgero.core.model.v1.match.builders.string.StringTypeBuilder;
+import com.sigmundgranaas.forgero.core.model.match.PredicateFactory;
+import com.sigmundgranaas.forgero.core.model.match.builders.string.StringIdentifierBuilder;
+import com.sigmundgranaas.forgero.core.model.match.builders.string.StringModelBuilder;
+import com.sigmundgranaas.forgero.core.model.match.builders.string.StringNameBuilder;
+import com.sigmundgranaas.forgero.core.model.match.builders.string.StringSlotBuilder;
+import com.sigmundgranaas.forgero.core.model.match.builders.string.StringTypeBuilder;
 import com.sigmundgranaas.forgero.core.property.active.ActivePropertyRegistry;
 import com.sigmundgranaas.forgero.core.property.active.VeinBreaking;
 import com.sigmundgranaas.forgero.core.registry.SoulLevelPropertyRegistry;
