+++ fabric/minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/match/DamagePercentagePredicate.java
@@ -5,8 +5,8 @@ import static com.sigmundgranaas.forgero.minecraft.common.match.MinecraftContext
 import java.util.Optional;
 
 import com.google.gson.JsonElement;
-import com.sigmundgranaas.forgero.core.model.v1.match.builders.ElementParser;
-import com.sigmundgranaas.forgero.core.model.v1.match.builders.PredicateBuilder;
+import com.sigmundgranaas.forgero.core.model.match.builders.ElementParser;
+import com.sigmundgranaas.forgero.core.model.match.builders.PredicateBuilder;
 import com.sigmundgranaas.forgero.core.util.match.MatchContext;
 import com.sigmundgranaas.forgero.core.util.match.Matchable;
