+++ forgero-core/src/main/java/com/sigmundgranaas/forgero/core/model/ModelConverter.java
@@ -14,8 +14,8 @@ import com.google.common.collect.ImmutableList;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
-import com.sigmundgranaas.forgero.core.model.v1.match.PredicateFactory;
-import com.sigmundgranaas.forgero.core.model.v1.match.PredicateMatcher;
+import com.sigmundgranaas.forgero.core.model.match.PredicateFactory;
+import com.sigmundgranaas.forgero.core.model.match.PredicateMatcher;
 import com.sigmundgranaas.forgero.core.resource.data.v2.data.DataResource;
 import com.sigmundgranaas.forgero.core.resource.data.v2.data.ModelData;
 import com.sigmundgranaas.forgero.core.resource.data.v2.data.ModelEntryData;
