+++ forgero-core/src/main/java/com/sigmundgranaas/forgero/core/model/match/builders/string/StringTypeBuilder.java
@@ -1,11 +1,11 @@
-package com.sigmundgranaas.forgero.core.model.v1.match.builders.string;
+package com.sigmundgranaas.forgero.core.model.match.builders.string;
 
 import java.util.Optional;
 
 import com.google.gson.JsonElement;
-import com.sigmundgranaas.forgero.core.model.v1.match.builders.ElementParser;
-import com.sigmundgranaas.forgero.core.model.v1.match.builders.PredicateBuilder;
-import com.sigmundgranaas.forgero.core.model.v1.match.predicate.TypePredicate;
+import com.sigmundgranaas.forgero.core.model.match.builders.ElementParser;
+import com.sigmundgranaas.forgero.core.model.match.builders.PredicateBuilder;
+import com.sigmundgranaas.forgero.core.model.match.predicate.TypePredicate;
 import com.sigmundgranaas.forgero.core.type.Type;
 import com.sigmundgranaas.forgero.core.util.match.Matchable;
