+++ forgero-core/src/main/java/com/sigmundgranaas/forgero/core/model/match/predicate/SlotPredicate.java
@@ -1,4 +1,4 @@
-package com.sigmundgranaas.forgero.core.model.v1.match.predicate;
+package com.sigmundgranaas.forgero.core.model.match.predicate;
 
 import com.sigmundgranaas.forgero.core.state.Slot;
 import com.sigmundgranaas.forgero.core.util.match.MatchContext;
